Use getqflist() on QuickFixCmdPre to decide which list to open
Some plugins - for example, tpope's fugitive - trigger custom
QuickFixCmdPost events using :doautocmd. This causes some commands - for
example, :Glgrep - to trigger inappropriate calls to qf#OpenQuickfix
when they should be triggering qf#OpenLoclist, due to the way qf.vim
determines which list to open (by command name).

To solve this, we can cache the value of getqflist() on QuickFixCmdPre,
and check it again on QuickFixCmdPost. If the value has changed, we know
to call qf#OpenQuickfix. Otherwise, we call qf#OpenLoclist. This should
allow for seamless integration with these sorts of plugin commands,
without having to know or care about their names.