Support count for <Plug>(qf_qf_previous) and friends
A count was ignored for <Plug>(qf_qf_previous), <Plug>(qf_qf_next),
<Plug>(qf_loc_previous), and <Plug>(qf_loc_next). This is a rather
simple implementation, and thus may not yet behave entirely as expected.

The count is only passed on to `:[cl]previous` and `:[cl]next`, but it
is ignored when we wrap using `:[cl]first` and `:[cl]last`. So
`100<Plug>(qf_qf_previous)` while at the top of the list will only jump
to the end, not to the 100-to-last item in the list.

Also, `:[cl]previous` and `:[cl]next` do not throw when the count
is greater than the number of items moved, e.g. `:3previous` while on
the 2nd entry in the list will move to the 1st entry, so it doesn't wrap
in that case.

I think both of these limitations are acceptable, but I'll gladly
implement something a bit more complex to handle these cases. Wanted to
check with you first, though, before doing unnecessary work.