Explicitly specify commands that trigger opening the lists. (#93)
The previous code was simpler, but assumed that every time
QuickFixCmdPost is fired, it's because of built in commands. However, I
ran into an issue where the vim-spellcheck plugin manually fires a
QuickFixCmdPost event with the name 'lspell', and it was recognized by
vim-qf as a trigger to open the location list, which was annoying.