Fix issue #76 (#115)
Improve the logic distinguishing the location list window from the
quickfix list window by inspecting the 'loclist' key in the output of
"getwininfo()".

When invoking "qf#SetList()" to clear the location list, bail out right
before ":lclose|lwindow".  Rationale: ":lwindow" will necessarily fail,
causing the subsequent invocation of "qf#SetList()" to unconditionally
use the quickfix list window.  We don't want that.  We want to stay in
the location list window if we're already in there.

Co-authored-by: lacygoill <lacygoill@lacygoill.me>