Fix `g:qf_shorten_path` behaviour
The docs state that the default for `g:qf_shorten_path` is 1:

https://github.com/romainl/vim-qf/blob/7e65325651ff5a0b06af8df3980d2ee54cf10e14/doc/qf.txt#L396-L398

But this isn't the case as it was defaulted to `0` in qf.vim.

Additionally, I believe the intention of this call:

```
call setqflist([], 'r', qf_list_title)
```

is to clear the qflist whilst retaining the title. However, it doesn't
actually clear the qflist because the first argument is ignored:

```
setqflist({list} [, {action} [, {what}]])               setqflist()
                Create or replace or add to the quickfix list.

                If the optional {what} dictionary argument is supplied, then
                only the items listed in {what} are set. The first {list}
                argument is ignored.  See below for the supported items in
                {what}.
```

This meant that when `g:qf_shorten_path` was set to > 0, the list with
shortened paths was being appended to the list with full paths,
resulting in duplicates.